Update 1.1
✔Fixed:
Fixed a bug where the player would get stuck in an infinite loop trying to increase the level.
Fixed a bug where the added levels and money were added with a defeat. (Now it will be added a tie)
Fixed an error trying to compile the project in netbeans (Broken references).

🆕Added:
Added a confirm pop up when you try to add money to your account.
Added text "version" in the auth-token screen.

❌Removed:
Removed cmd displaying unecesary information.
Some comments deleted.